# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ThreatCheck API",
    "description": "A RESTful API for scanning files using ClamAV antivirus engine with PostgreSQL logging. This API provides endpoints to scan single or multiple files, calculate file hashes (MD5, SHA1, SHA256), and maintain a complete audit log of all scan activities.",
    "version": "1.0.0",
    "contact": {
      "name": "ThreatCheck API Support",
      "email": "support@threatcheck.com"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.threatcheck.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/scan/file": {
      "post": {
        "tags": ["Scan"],
        "summary": "Scan a single file",
        "description": "Upload and scan a single file for threats. The file will be scanned using ClamAV and its MD5, SHA1, and SHA256 hashes will be calculated. Results are logged to the database.",
        "operationId": "scanSingleFile",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "The file to scan (max 100MB)"
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File scanned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleScanResponse"
                },
                "example": {
                  "success": true,
                  "scanId": 123,
                  "filename": "document.pdf",
                  "fileSize": 1024,
                  "hashes": {
                    "md5": "5d41402abc4b2a76b9719d911017c592",
                    "sha1": "aaf4c61ddcc5e8a2dabede0f3b482cd9aea9434d",
                    "sha256": "2c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7ae"
                  },
                  "scanResult": {
                    "isClean": true,
                    "isInfected": false,
                    "threats": [],
                    "scanTime": "2025-07-02T10:30:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - no file uploaded or file too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "noFile": {
                    "summary": "No file uploaded",
                    "value": {
                      "success": false,
                      "error": "No file uploaded"
                    }
                  },
                  "fileTooLarge": {
                    "summary": "File too large",
                    "value": {
                      "success": false,
                      "error": "File too large",
                      "message": "File size exceeds 100MB limit"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error during scan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Internal server error during file scan",
                  "message": "ClamAV service unavailable"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/scan/history": {
      "get": {
        "tags": ["History"],
        "summary": "Get scan history",
        "description": "Retrieve the history of all file scans from the database with pagination support",
        "operationId": "getScanHistory",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Number of records to return",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 1000
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of records to skip",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scan history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanHistoryResponse"
                },
                "example": {
                  "success": true,
                  "count": 2,
                  "limit": 50,
                  "offset": 0,
                  "data": [
                    {
                      "id": 123,
                      "filename": "document.pdf",
                      "file_size": 1024,
                      "md5_hash": "5d41402abc4b2a76b9719d911017c592",
                      "sha1_hash": "aaf4c61ddcc5e8a2dabede0f3b482cd9aea9434d",
                      "sha256_hash": "2c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7ae",
                      "scan_result": "CLEAN",
                      "threats_found": [],
                      "scan_time": "2025-07-02T10:30:00.000Z",
                      "client_ip": "192.168.1.100",
                      "user_agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36"
                    },
                    {
                      "id": 124,
                      "filename": "malicious.exe",
                      "file_size": 2048,
                      "md5_hash": "44d88612fea8a8f36de82e1278abb02f",
                      "sha1_hash": "3395856ce81f2b7382dee72602f798b642f14140",
                      "sha256_hash": "275a021bbfb6489e54d471899f7db9d1663fc695ec2fe2a2c4538aabf651fd0f",
                      "scan_result": "INFECTED",
                      "threats_found": ["Win.Test.EICAR_HDB-1"],
                      "scan_time": "2025-07-02T10:29:00.000Z",
                      "client_ip": "192.168.1.101",
                      "user_agent": "curl/7.68.0"
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Error retrieving scan history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Failed to fetch scan history",
                  "message": "Database connection error"
                }
              }
            }
          }
        }
      }
    },
    "/api/signatures/info": {
      "get": {
        "tags": ["Signatures"],
        "summary": "Get ClamAV signature information",
        "description": "Retrieve current ClamAV signature database information including version, signature counts, and update history",
        "operationId": "getSignatureInfo",
        "responses": {
          "200": {
            "description": "Signature information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignatureInfoResponse"
                },
                "example": {
                  "success": true,
                  "timestamp": "2025-07-03T10:30:00.000Z",
                  "current": {
                    "version": "ClamAV 1.0.0/26000/Mon Jul  1 10:30:00 2024",
                    "databases": [
                      {
                        "name": "main.cvd",
                        "signatures": 6647427,
                        "lastUpdate": "Mon Jul  1 10:30:00 2024"
                      },
                      {
                        "name": "daily.cvd",
                        "signatures": 2038936,
                        "lastUpdate": "Mon Jul  1 10:30:00 2024"
                      }
                    ],
                    "lastUpdate": "Mon Jul  1 10:30:00 2024",
                    "totalSignatures": 8686363
                  },
                  "updateHistory": []
                }
              }
            }
          },
          "500": {
            "description": "Error retrieving signature information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Failed to fetch signature information",
                  "message": "ClamAV service unavailable"
                }
              }
            }
          }
        }
      }
    },
    "/api/signatures/update": {
      "post": {
        "tags": ["Signatures"],
        "summary": "Update ClamAV signatures",
        "description": "Trigger an update of ClamAV signature databases and log the update status",
        "operationId": "updateSignatures",
        "responses": {
          "200": {
            "description": "Signature update completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignatureUpdateResponse"
                },
                "example": {
                  "success": true,
                  "timestamp": "2025-07-03T10:30:00.000Z",
                  "updateResult": {
                    "updated": true,
                    "databases": ["main.cvd", "daily.cvd", "bytecode.cvd"],
                    "message": "Signature update completed successfully"
                  },
                  "before": {
                    "totalSignatures": 8686363
                  },
                  "after": {
                    "totalSignatures": 8686500
                  },
                  "message": "Signature update completed successfully"
                }
              }
            }
          },
          "500": {
            "description": "Error updating signatures",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Failed to update signatures",
                  "message": "ClamAV update service unavailable"
                }
              }
            }
          }
        }
      }
    },
    "/api/signatures/history": {
      "get": {
        "tags": ["Signatures"],
        "summary": "Get signature update history",
        "description": "Retrieve the history of ClamAV signature database updates with pagination support",
        "operationId": "getSignatureHistory",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Number of records to return",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 1000
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of records to skip",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Signature update history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignatureHistoryResponse"
                },
                "example": {
                  "success": true,
                  "count": 2,
                  "limit": 50,
                  "offset": 0,
                  "data": [
                    {
                      "id": 1,
                      "database_name": "main.cvd",
                      "version": "ClamAV 1.0.0/26000/Mon Jul  1 10:30:00 2024",
                      "signatures_count": 6647427,
                      "last_updated": "2025-07-03T10:30:00.000Z",
                      "update_status": "SUCCESS",
                      "update_details": "Signature update completed successfully",
                      "file_size": null
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Error retrieving signature update history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Failed to fetch signature update history",
                  "message": "Database connection error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "SingleScanResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [true],
            "description": "Indicates successful operation"
          },
          "scanId": {
            "type": "integer",
            "description": "Unique identifier for this scan record"
          },
          "filename": {
            "type": "string",
            "description": "Original filename of the uploaded file"
          },
          "fileSize": {
            "type": "integer",
            "description": "Size of the file in bytes"
          },
          "hashes": {
            "$ref": "#/components/schemas/FileHashes"
          },
          "scanResult": {
            "$ref": "#/components/schemas/ScanResult"
          }
        },
        "required": ["success", "scanId", "filename", "fileSize", "hashes", "scanResult"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [false],
            "description": "Indicates failed operation"
          },
          "error": {
            "type": "string",
            "description": "Error description"
          },
          "message": {
            "type": "string",
            "description": "Detailed error message"
          }
        },
        "required": ["success", "error"]
      },
...
    }
  },
  "security": [],
...
}